Changed login page error message and also handles the error variable problem
So if I  failed login it shows an error message and then if I go to the signup form and create an account successfully it redirects me to the login page. But on login page, it was showing me the old error message. Solve this error by unseting failed_attempts after successful sign up.